Update packages (#246)
* Update Karate, Quarkus and Graal and fix issue with passing JSON objects.
* Add method to output model as triples
* Change method of converting JSON to a string

+++ src/main/java/org/solid/testharness/config/TestSubject.java
@@ -259,7 +259,9 @@ public class TestSubject {
 
     public void tearDownServer() {
         try {
-            testRunContainer.delete();
+            if (testRunContainer != null) {
+                testRunContainer.delete();
+            }
         } catch (Exception e) {
             // log failure but continue to report results
             logger.error("Failed to delete the test containers", e);
